MSVC improvements (#1)
* Some changes to be able to compile serializer_tests for MSVC

* Addressed various issues compiling for MSVC

* Additional test compatibility for MSVC

* Added a template specifier back where needed

+++ include/nop/base/vector.h
@@ -59,7 +59,7 @@ struct Encoding<std::vector<T, Allocator>, EnableIfNotIntegral<T>>
   static constexpr std::size_t Size(const Type& value) {
     return BaseEncodingSize(Prefix(value)) +
            Encoding<SizeType>::Size(value.size()) +
-           std::accumulate(value.cbegin(), value.cend(), 0U,
+           std::accumulate(value.cbegin(), value.cend(), static_cast<size_t>(0),
                            [](const std::size_t& sum, const T& element) {
                              return sum + Encoding<T>::Size(element);
                            });
